% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmOut_function.R
\name{lmOut}
\alias{lmOut}
\title{Regression summary output}
\usage{
lmOut(res, ndigit = 3, writecsv = FALSE, file = "")
}
\arguments{
\item{res}{regression model.}

\item{ndigit}{number of rounding digits.}

\item{writecsv}{write to csv flag. Defaults to FALSE}

\item{file}{output CSV file name. Defaults to empty string}
}
\description{
This function collects coefficients, t-stats, p-values, f-stats, etc from a regression summary and writes output to csv or returns dataframe.
}
\examples{
lmOut(myres)
lmOut(myres, writecsv=TRUE, file="my-results.csv")
write.xlsx(lmOut(myres), file=filepath, sheetName="Summary", row.names=FALSE)
}
\keyword{lmOut}
